// src/features/home/components/ContactSection/schemas/contact.schema.ts
import { z } from "zod";

export const contactSchema = z.object({
    name: z.string().min(2, "El nombre debe tener al menos 2 caracteres."),
    email: z.string().email("Por favor, ingresa un correo electrónico válido."),
    message: z.string().min(10, "El mensaje debe tener al menos 10 caracteres para entender tu solicitud.")
});

export type ContactFormValues = z.infer<typeof contactSchema>;